requirements.txt
Flask==3.0.0
Flask-Session==0.5.0
Flask-Login==0.6.3
Werkzeug==3.0.1
Flask-Compress==1.13
Pillow==10.1.0
cairosvg==2.7.1